Test all flex upgrades.

+++ test/test_samples.rb
@@ -8,8 +8,8 @@ require_relative '../resources/buildstock'
 
 class IntegrationWorkflowTest < MiniTest::Test
   def before_setup
-    @project_dir_baseline = { 'project_testing' => 10, 'project_national' => 30 }
-    @project_dir_upgrades = { 'project_testing' => 1, 'project_national' => 1 }
+    @project_dir_baseline = { 'project_testing' => 0, 'project_national' => 10 }
+    @project_dir_upgrades = { 'project_testing' => 0, 'project_national' => 10 }
 
     @outfile = File.join('..', 'test', 'test_samples_osw', 'buildstock.csv')
     @top_dir = File.absolute_path(File.join(File.dirname(__FILE__), 'test_samples_osw'))
@@ -19,10 +19,10 @@ class IntegrationWorkflowTest < MiniTest::Test
 
   def after_teardown
     FileUtils.rm_rf(@lib_dir) if File.exist?(@lib_dir)
-    FileUtils.rm_rf(File.join(@top_dir, 'run'))
-    FileUtils.rm_rf(File.join(@top_dir, 'reports'))
-    FileUtils.rm_rf(File.join(@top_dir, 'generated_files'))
-    FileUtils.rm(File.join(@top_dir, 'buildstock.csv'))
+    FileUtils.rm_rf(File.join(@top_dir, 'run')) if File.exist?(File.join(@top_dir, 'run'))
+    FileUtils.rm_rf(File.join(@top_dir, 'reports')) if File.exist?(File.join(@top_dir, 'reports'))
+    FileUtils.rm_rf(File.join(@top_dir, 'generated_files')) if File.exist?(File.join(@top_dir, 'generated_files'))
+    FileUtils.rm(File.join(@top_dir, 'buildstock.csv')) if File.exist?(File.join(@top_dir, 'buildstock.csv'))
   end
 
   def test_baseline
@@ -31,6 +31,8 @@ class IntegrationWorkflowTest < MiniTest::Test
 
     all_results = []
     @project_dir_baseline.each do |project_dir, num_samples|
+      next unless num_samples > 0
+
       samples_osw(scenario_dir, project_dir, num_samples, all_results)
     end
 
@@ -50,11 +52,13 @@ class IntegrationWorkflowTest < MiniTest::Test
   end
 
   def test_upgrades
-    scenario_dir = File.join(@top_dir, 'upgrades')
+    scenario_dir = File.join(@top_dir, 'upgrades-flex')
     Dir.mkdir(scenario_dir) unless File.exist?(scenario_dir)
-    
+
     all_results = []
     @project_dir_upgrades.each do |project_dir, num_samples|
+      next unless num_samples > 0
+
       samples_osw(scenario_dir, project_dir, num_samples, all_results)
     end
